+++ packages/common/src/constants.ts
@@ -581,6 +581,7 @@ export const ALCHEMY_ETHEREUM_MAINNET_API_KEY =
 export const AVATAR_BASE_URL = "https://swr.xnfts.dev/avatars";
 export const BACKEND_API_URL = "https://backpack-api.xnfts.dev";
 export const REALTIME_API_URL = "https://backend-ws.xnfts.dev";
+export const REALTIME_WS_URL = "wss://backend-ws.xnfts.dev";
 export const MESSAGING_COMMUNICATION_PUSH = "MESSAGING_COMMUNICATION_PUSH";
 export const MESSAGING_COMMUNICATION_FETCH = "MESSAGING_COMMUNICATION_FETCH";
 export const MESSAGING_COMMUNICATION_FETCH_RESPONSE =
