--- packages/react-common/src/config.ts
@@ -1,1 +1,0 @@
-export const SERVER_URL = "wss://backend-ws.xnfts.dev";
